rework next/link flag for Next.js 12 and 13
