+++ code/frameworks/nextjs/src/preset.ts
@@ -1,6 +1,5 @@
 // https://storybook.js.org/docs/react/addons/writing-presets
 import { dirname, join } from 'path';
-import semver from 'semver';
 import type { Options, PresetProperty } from '@storybook/types';
 import type { TransformOptions } from '@babel/core';
 import { configureConfig } from './config/webpack';
@@ -10,7 +9,7 @@ import { configureRouting } from './routing/webpack';
 import { configureStyledJsx } from './styledJsx/webpack';
 import { configureStyledJsxTransforms } from './styledJsx/babel';
 import { configureImages } from './images/webpack';
-import { configureRuntimeNextjsVersionResolution, getNextjsVersion } from './utils';
+import { configureRuntimeNextjsVersionResolution } from './utils';
 import type { FrameworkOptions, StorybookConfig } from './types';
 import { configureTypescript } from './config/babel';
 
@@ -18,8 +17,6 @@ export const addons: PresetProperty<'addons', StorybookConfig> = [
   dirname(require.resolve(join('@storybook/preset-react-webpack', 'package.json'))),
   dirname(require.resolve(join('@storybook/builder-webpack5', 'package.json'))),
 ];
-
-const version = getNextjsVersion();
 
 const defaultFrameworkOptions: FrameworkOptions = {};
 
@@ -70,17 +67,6 @@ export const config: StorybookConfig['previewAnnotations'] = (entry = []) => [
   ...entry,
   require.resolve('@storybook/nextjs/preview.js'),
 ];
-
-export const env = (envConfig: PresetProperty<'env', StorybookConfig>) => {
-  return {
-    ...envConfig,
-    // Can be removed if https://github.com/vercel/next.js/issues/42621 is resolved
-    ...(semver.gte(version, '13.0.0') && {
-      // TODO: This should also respect `newNextLinkBehavior`: https://github.com/vercel/next.js/blob/07d3da102dfef65be9c13fd4b754a12eda7eded1/packages/next/server/config-shared.ts#L88
-      __NEXT_NEW_LINK_BEHAVIOR: 'true',
-    }),
-  };
-};
 
 // Not even sb init - automigrate - running dev
 // You're using a version of Nextjs prior to v10, which is unsupported by this framework.
