+++ code/frameworks/nextjs/src/utils.ts
@@ -1,6 +1,11 @@
 import path from 'path';
-import type { Configuration as WebpackConfig } from 'webpack';
 import { DefinePlugin } from 'webpack';
+import { PHASE_DEVELOPMENT_SERVER } from 'next/constants';
+import findUp from 'find-up';
+import { pathExists } from 'fs-extra';
+import dedent from 'ts-dedent';
+import type { Configuration as WebpackConfig } from 'webpack';
+import type { NextConfig } from 'next';
 
 export const configureRuntimeNextjsVersionResolution = (baseConfig: WebpackConfig): void => {
   baseConfig.plugins?.push(
@@ -11,6 +16,54 @@ export const configureRuntimeNextjsVersionResolution = (baseConfig: WebpackConfi
 };
 
 export const getNextjsVersion = (): string => require(scopedResolve('next/package.json')).version;
+
+const findNextConfigFile = async (configDir: string) => {
+  const supportedExtensions = ['mjs', 'js'];
+  return supportedExtensions.reduce<Promise<undefined | string>>(
+    async (acc, ext: string | undefined) => {
+      const resolved = await acc;
+      if (!resolved) {
+        acc = findUp(`next.config.${ext}`, { cwd: configDir });
+      }
+
+      return acc;
+    },
+    Promise.resolve(undefined)
+  );
+};
+
+export const resolveNextConfig = async ({
+  baseConfig = {},
+  nextConfigPath,
+  configDir,
+}: {
+  baseConfig?: WebpackConfig;
+  nextConfigPath?: string;
+  configDir: string;
+}): Promise<NextConfig> => {
+  const nextConfigFile = nextConfigPath || (await findNextConfigFile(configDir));
+
+  if (!nextConfigFile || (await pathExists(nextConfigFile)) === false) {
+    throw new Error(
+      dedent`
+        Could not find or resolve your Next config file. Please provide the next config file path as a framework option.
+
+        More info: https://github.com/storybookjs/storybook/blob/next/code/frameworks/nextjs/README.md#options
+      `
+    );
+  }
+
+  const nextConfigExport = await import(nextConfigFile);
+
+  const nextConfig =
+    typeof nextConfigExport === 'function'
+      ? nextConfigExport(PHASE_DEVELOPMENT_SERVER, {
+          defaultConfig: baseConfig,
+        })
+      : nextConfigExport;
+
+  return nextConfig.default || nextConfig;
+};
 
 // This is to help the addon in development
 // Without it, webpack resolves packages in its node_modules instead of the example's node_modules
